Added check for Children

+++ src/bundle/Command/DatabaseHealthCheckCommand.php
@@ -8,22 +8,30 @@ use eZ\Publish\API\Repository\ContentService;
 use eZ\Publish\API\Repository\PermissionResolver;
 use eZ\Publish\API\Repository\Repository;
 use eZ\Publish\Core\MVC\Symfony\SiteAccess;
-use eZ\Publish\SPI\Persistence\Content\Handler;
+use eZ\Publish\SPI\Persistence\Content\Handler as ContentHandler;
+use eZ\Publish\SPI\Persistence\Content\Location\Handler as LocationHandler;
 use MateuszBieniek\EzPlatformDatabaseHealthChecker\Dto\CorruptedAttribute;
 use MateuszBieniek\EzPlatformDatabaseHealthChecker\Dto\CorruptedContent;
 use MateuszBieniek\EzPlatformDatabaseHealthChecker\Persistence\Legacy\Content\Gateway\GatewayInterface as ContentGateway;
+use eZ\Publish\Core\Persistence\Legacy\Content\Gateway as EzContentGateway;
 use Symfony\Component\Console\Command\Command;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Input\InputOption;
 use Symfony\Component\Console\Output\OutputInterface;
+use Symfony\Component\Console\Question\ChoiceQuestion;
 use Symfony\Component\Console\Style\SymfonyStyle;
 
 class DatabaseHealthCheckCommand extends Command
 {
     private const CONTENT_LIMIT = 100;
+    private const ACTION_SKIP = 'skip';
+    private const ACTION_DELETE = 'delete';
 
     /** @var ContentGateway */
     private $contentGateway;
+
+    /** @var EzContentGateway */
+    private $ezContentGateway;
 
     /** @var \eZ\Publish\API\Repository\ContentService */
     private $contentService;
@@ -37,25 +45,32 @@ class DatabaseHealthCheckCommand extends Command
     /** @var \eZ\Publish\API\Repository\Repository */
     private $repository;
 
-    /** @var \eZ\Publish\SPI\Persistence\Content\Handler */
+    /** @var ContentHandler */
     private $contentHandler;
+
+    /** @var LocationHandler */
+    private $locationHandler;
 
     /** @var \Symfony\Component\Console\Style\SymfonyStyle */
     private $io;
 
     public function __construct(
         ContentGateway $contentGateway,
+        EzContentGateway $ezContentGateway,
         ContentService $contentService,
         SiteAccess $siteAccess,
         PermissionResolver $permissionResolver,
-        Handler $contentHandler,
+        ContentHandler $contentHandler,
+        LocationHandler $locationHandler,
         Repository $repository
     ) {
         $this->contentGateway = $contentGateway;
+        $this->ezContentGateway = $ezContentGateway;
         $this->contentService = $contentService;
         $this->siteAccess = $siteAccess;
         $this->permissionResolver = $permissionResolver;
         $this->contentHandler = $contentHandler;
+        $this->locationHanlder = $locationHandler;
         $this->repository = $repository;
 
         parent::__construct();
@@ -115,14 +130,14 @@ EOT
             $this->smokeTest();
         }
 
-        $this->checkContentWithoutAttributes();
+        $this->checkContentWithoutAttributes($input, $output);
         $this->checkContentWithoutVersions();
         $this->checkDuplicatedAttributes();
 
         $this->io->success('Done');
     }
 
-    private function checkContentWithoutAttributes()
+    private function checkContentWithoutAttributes(InputInterface $input, OutputInterface $output)
     {
         $this->io->section('Searching for Content without attributes.');
         $this->io->text('It may take some time. Please wait...');
@@ -154,8 +169,37 @@ EOT
         if ($this->io->confirm('Do you want to proceed?', false)) {
             $progressBar = $this->io->createProgressBar($contentWithoutAttributesCount);
 
-            foreach ($contentWithoutAttributes  as $corruptedContent) {
-                $this->contentHandler->deleteContent($corruptedContent->id);
+            $helper = $this->getHelper('question');
+            $corruptedParentLocationQuestion = new ChoiceQuestion(
+                'Please select what you want to do with this Content:',
+                [self::ACTION_SKIP, self::ACTION_DELETE],
+                self::ACTION_SKIP
+            );
+            $corruptedParentLocationQuestion->setErrorMessage('Option %s is invalid.');
+
+            foreach ($contentWithoutAttributes as $corruptedContent) {
+                $corruptedLocations = $this->ezContentGateway->getAllLocationIds($corruptedContent->id);
+                $action = self::ACTION_DELETE;
+
+                foreach ($corruptedLocations as $corruptedLocation) {
+                    if ($this->hasLocationChildren($corruptedLocation)) {
+                        $this->io->warning(
+                            sprintf('Location %d of Content %d has children, which will be removed as well.',
+                                $corruptedLocation,
+                                $corruptedContent->id
+                            )
+                        );
+                        $action = $helper->ask($input, $output, $corruptedParentLocationQuestion);
+
+                        if ($action === self::ACTION_SKIP) {
+                            break;
+                        }
+                    }
+                }
+
+                if($action === self::ACTION_DELETE) {
+                    $this->contentHandler->deleteContent($corruptedContent->id);
+                }
                 $progressBar->advance(1);
             }
 
@@ -196,6 +240,7 @@ EOT
             $progressBar = $this->io->createProgressBar($contentWithoutVersionsCount);
 
             foreach ($contentWithoutVersions  as $corruptedContent) {
+                $this->contentHandler->
                 $this->contentHandler->deleteContent($corruptedContent->id);
                 $progressBar->advance(1);
             }
@@ -305,5 +350,14 @@ EOT
             ['Content ID', 'Error message'],
             $erroredContents
         );
+    }
+
+    private function hasLocationChildren(int $locationId): bool
+    {
+        if(!empty($this->locationHandler->loadSubtreeIds($locationId))) {
+            return true;
+        }
+
+        return false;
     }
 }
